pyproject.toml
[tool.poetry]
name = "Redis Python Tutorial"
version = "0.1.0"
description = ""
authors = ["Todd Birchard <toddbirchard@gmail.com>"]
maintainers = ["Todd Birchard <toddbirchard@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://hackersandslackers.com/redis-py-python/"
repository = "https://github.com/hackersandslackers/redis-python-tutorial/"
documentation = "https://hackersandslackers.com/redis-py-python/"
keywords = [
    "Redis",
    "Cache",
    "Queue",
    "Memory",
    "NoSQL"
]

[tool.poetry.dependencies]
python = "3.9"
redis = "*"
loguru = "*"
python-dotenv = "*"

[tool.poetry.dev-dependencies]
pytest = "*"
coverage = "*"
mock = "*"
mypy = "*"
black = "*"
isort = "*"
flake8 = "*"
pylint = "*"

[build-system]
requires = ["poetry>=1.1.12"]
build-backend = "poetry.masonry.api"

[tool.poetry.scripts]
run = "main:init_script"

[tool.poetry.urls]
issues = "https://github.com/hackersandslackers/redis-python-tutorial/issues"

[tool.black]
line-length = 100
